Properly support browsers
Before, I don't think you could actually use this package because it never had a global name in the `iife` case...

So this gives a global name to the `iife` and also adds an `esm` target for anyone using modules.

I added a couple of html files that act as smoke tests for these two cases too.