Handle newlines correctly
Before, we were querying for only `string_fragments` but those fragments are not actually the full string. If your string has a `\n` in it, then your `string` has 2 `string_fragments` in it.

So this PR selects the `string` instead of the `string_fragment`... but that has a downside that it includes the quotes... so we need to exclude the quotes.