Handle newlines in XLFs better
Newlines being treated as newlines seemed to mess up the XML parser.
This change now turns them into their charcodes `10` and `13` so that
they can be parsed correctly.

With that said, one bug I noticed in the XML parser is that if a value
only contains newline characters, then it treats it as an empty string.
I will open an issue about this in the xml2js repo.

Fixes #63