Better handling of grammar and parser (#72)
* Better handling of grammar and parser

In the VS Code repo, I was seeing an issue where the parser and grammar were being loaded multiple times

And in some cases, because of my if not checking if the grammar was defined, it would totally fail.

This change ensures that both grammar and parser will be defined.

* bump