Detect built-in (aka language pack) extensions
Extensions that are a part of a language pack have a different schema than extensions that bundle their strings with the extension.

This change detects the difference and initializes the bundle accordingly. This is a temporary solution for #84 until a larger change can be made. Either:

* Reworking the structure of language packs to separate bundle and package entrypoints
* Having the client cache the language pack extensions in a way such that they align with extensions that ship their own strings
* Changing the schema of the bundles in extensions that ship their own strings to align with language pack extensions

There's a lot to research there, so in the meantime, this fix unblocks us.

Fixes #84